fix: fan out health checks into individual jobs per service
The scheduled CheckHealthJob was running all service health checks
synchronously in a single loop, causing jobs to run for 65+ hours.
Now it enqueues a separate CheckServiceHealthJob for each service,
so checks run in parallel and a slow/hanging kubectl call only
blocks that one service's check.
